Route Outlook email content through the unified add menu
