+++ office-addin/src/locales/pl/messages.po
@@ -370,81 +370,81 @@ msgstr "Szkic jest zbyt duży, aby wstępnie wypełnić odpowiedź. Użyj Kopiuj
 
 #. js-lingui-explicit-id
 #: src/components/AddinFileSourceSelector.tsx
-msgid "officeAddin.fileSource.addFiles"
-msgstr "Dodaj pliki"
+#~ msgid "officeAddin.fileSource.addFiles"
+#~ msgstr "Dodaj pliki"
 
 #. js-lingui-explicit-id
-#: src/components/AddinFileSourceSelector.tsx
+#: src/components/AddinChatAddMenuExtraContent.tsx
 msgid "officeAddin.fileSource.alreadyAdded"
 msgstr "Już dodano"
 
 #. js-lingui-explicit-id
-#: src/components/AddinFileSourceSelector.tsx
+#: src/components/AddinChatAddMenuExtraContent.tsx
 msgid "officeAddin.fileSource.attachmentFallback"
 msgstr "Załącznik"
 
 #. js-lingui-explicit-id
-#: src/components/AddinFileSourceSelector.tsx
+#: src/components/AddinChatAddMenuExtraContent.tsx
 msgid "officeAddin.fileSource.cloudAttachmentUnsupported"
 msgstr "Załącznika w chmurze nie można przesłać z Outlooka"
 
 #. js-lingui-explicit-id
-#: src/components/AddinFileSourceSelector.tsx
+#: src/components/AddinChatAddMenuExtraContent.tsx
 msgid "officeAddin.fileSource.emailContent"
 msgstr "Treść wiadomości"
 
 #. js-lingui-explicit-id
-#: src/components/AddinFileSourceSelector.tsx
+#: src/components/AddinChatAddMenuExtraContent.tsx
 msgid "officeAddin.fileSource.emailThread"
 msgstr "Wątek wiadomości"
 
 #. js-lingui-explicit-id
-#: src/components/AddinFileSourceSelector.tsx
+#: src/components/AddinChatAddMenuExtraContent.tsx
 msgid "officeAddin.fileSource.emailThreadLoadError"
 msgstr "Nie udało się załadować tej konwersacji z serwera. Niektóre wiadomości lub załączniki mogą być niedostępne — spróbuj ponownie otworzyć wiadomość."
 
 #. js-lingui-explicit-id
 #: src/components/AddinFileSourceSelector.tsx
-msgid "officeAddin.fileSource.fromComputer"
-msgstr "Prześlij z komputera"
+#~ msgid "officeAddin.fileSource.fromComputer"
+#~ msgstr "Prześlij z komputera"
 
 #. js-lingui-explicit-id
 #: src/components/AddinFileSourceSelector.tsx
-msgid "officeAddin.fileSource.fromOneDrive"
-msgstr "Prześlij z Sharepoint"
+#~ msgid "officeAddin.fileSource.fromOneDrive"
+#~ msgstr "Prześlij z Sharepoint"
 
 #. js-lingui-explicit-id
 #: src/components/AddinFileSourceSelector.tsx
-msgid "officeAddin.fileSource.hide"
-msgstr "Ukryj"
+#~ msgid "officeAddin.fileSource.hide"
+#~ msgstr "Ukryj"
 
 #. js-lingui-explicit-id
 #: src/components/AddinFileSourceSelector.tsx
-msgid "officeAddin.fileSource.loading"
-msgstr "..."
+#~ msgid "officeAddin.fileSource.loading"
+#~ msgstr "..."
 
 #. js-lingui-explicit-id
-#: src/components/AddinFileSourceSelector.tsx
+#: src/components/AddinChatAddMenuExtraContent.tsx
 msgid "officeAddin.fileSource.loadingAttachments"
 msgstr "Ładowanie załączników..."
 
 #. js-lingui-explicit-id
-#: src/components/AddinFileSourceSelector.tsx
+#: src/components/AddinChatAddMenuExtraContent.tsx
 msgid "officeAddin.fileSource.loadingEmailThread"
 msgstr "Ładowanie wątku wiadomości..."
 
 #. js-lingui-explicit-id
 #: src/components/AddinFileSourceSelector.tsx
-msgid "officeAddin.fileSource.noEmailContent"
-msgstr "Dla tego elementu nie jest dostępna treść wiadomości."
+#~ msgid "officeAddin.fileSource.noEmailContent"
+#~ msgstr "Dla tego elementu nie jest dostępna treść wiadomości."
 
 #. js-lingui-explicit-id
 #: src/components/AddinFileSourceSelector.tsx
-msgid "officeAddin.fileSource.show"
-msgstr "Pokaż"
+#~ msgid "officeAddin.fileSource.show"
+#~ msgstr "Pokaż"
 
 #. js-lingui-explicit-id
-#: src/components/AddinFileSourceSelector.tsx
+#: src/components/AddinChatAddMenuExtraContent.tsx
 msgid "officeAddin.fileSource.updatingThread"
 msgstr "Aktualizowanie…"
